python/pyproject.toml
[project]
name = "carfac"
version = "0.0.1"
description = "Python implementation of Cascade of Asymmetric Resonators with Fast-Acting Compression (CARFAC)"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib~=3.9.1",
    "numpy~=2.1.3",
]

[project.optional-dependencies]
tf = [
    "tensorflow~=2.19.0",
    "tf-keras~=2.19.0",
]
jax = [
    "jax~=0.6.0",
]

[dependency-groups]
dev = [
    "absl-py~=2.2.2",
    "google-benchmark~=1.9.4",
    "parameterized~=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Tests and test data are included in distributions, but that is okay as we are
# only distributing this package as a Git repository (which will include
# everything when cloned anyway).
force-include = {"../test_data" = "carfac/test_data"}
packages = ["src/carfac"]
